# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Autoken Facturas v2",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/platform/settings": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "Get Settings",
        "operationId": "get_settings_api_v1_platform_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformSettingsOut"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "platform"
        ],
        "summary": "Set Settings",
        "operationId": "set_settings_api_v1_platform_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlatformSettingsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformSettingsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/platform/ocr-lab/settings": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "Get Lab Settings",
        "operationId": "get_lab_settings_api_v1_platform_ocr_lab_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OcrLabSettingsOut"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "platform"
        ],
        "summary": "Set Lab Settings",
        "operationId": "set_lab_settings_api_v1_platform_ocr_lab_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OcrLabSettingsPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OcrLabSettingsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/platform/ocr-lab/promote": {
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Promote Ocr Policy",
        "operationId": "promote_ocr_policy_api_v1_platform_ocr_lab_promote_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OcrPolicy"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OcrPolicyPromotionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OcrLabSettingsOut": {
        "properties": {
          "lab_visible": {
            "type": "boolean",
            "title": "Lab Visible"
          },
          "auto_benchmark_enabled": {
            "type": "boolean",
            "title": "Auto Benchmark Enabled"
          },
          "benchmark_engines": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Benchmark Engines"
          },
          "benchmark_variants": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Benchmark Variants"
          }
        },
        "type": "object",
        "required": [
          "lab_visible",
          "auto_benchmark_enabled",
          "benchmark_engines",
          "benchmark_variants"
        ],
        "title": "OcrLabSettingsOut"
      },
      "OcrLabSettingsPayload": {
        "properties": {
          "lab_visible": {
            "type": "boolean",
            "title": "Lab Visible"
          },
          "auto_benchmark_enabled": {
            "type": "boolean",
            "title": "Auto Benchmark Enabled"
          },
          "benchmark_engines": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Benchmark Engines"
          },
          "benchmark_variants": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Benchmark Variants"
          }
        },
        "type": "object",
        "required": [
          "lab_visible",
          "auto_benchmark_enabled",
          "benchmark_engines",
          "benchmark_variants"
        ],
        "title": "OcrLabSettingsPayload"
      },
      "OcrPolicy": {
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          },
          "primary_engine": {
            "type": "string",
            "minLength": 1,
            "title": "Primary Engine"
          },
          "primary_model": {
            "type": "string",
            "minLength": 1,
            "title": "Primary Model"
          },
          "fallback_enabled": {
            "type": "boolean",
            "title": "Fallback Enabled"
          },
          "fallback_engine": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Engine"
          },
          "fallback_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Model"
          },
          "consensus_mode": {
            "type": "string",
            "enum": [
              "primary_only",
              "per_field"
            ],
            "title": "Consensus Mode"
          }
        },
        "type": "object",
        "required": [
          "version",
          "primary_engine",
          "primary_model",
          "fallback_enabled",
          "consensus_mode"
        ],
        "title": "OcrPolicy",
        "description": "Pol\u00edtica versionada que decide qu\u00e9 motor puede usar producci\u00f3n."
      },
      "OcrPolicyPromotionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "old_policy": {
            "$ref": "#/components/schemas/OcrPolicy"
          },
          "new_policy": {
            "$ref": "#/components/schemas/OcrPolicy"
          },
          "actor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Actor Id"
          },
          "promoted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Promoted At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "old_policy",
          "new_policy",
          "actor_id",
          "promoted_at"
        ],
        "title": "OcrPolicyPromotionOut"
      },
...
      "PlatformSettingsIn": {
        "properties": {
          "ocr_experiment_enabled": {
            "type": "boolean",
            "title": "Ocr Experiment Enabled"
          }
        },
        "type": "object",
        "required": [
          "ocr_experiment_enabled"
        ],
        "title": "PlatformSettingsIn"
      },
      "PlatformSettingsOut": {
        "properties": {
          "ocr_experiment_enabled": {
            "type": "boolean",
            "title": "Ocr Experiment Enabled"
          }
        },
        "type": "object",
        "required": [
          "ocr_experiment_enabled"
        ],
        "title": "PlatformSettingsOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}